bpf: test: cover NAT LB with catch-all EGW policy
Reply traffic for an SNATed in-cluster nodeport connection should not be
captured by a catch-all EGW policy. Otherwise the packet reaches the client
without the source being RevDNATed back to the nodeport, and thus
resulting in a RST by the client.

This adds testing for the scenario that was fixed in the preceding commit.
