+++ bpf/tests/lib/egressgw.h
@@ -14,6 +14,8 @@
 #define EGRESS_IP2		IPV4(2, 3, 4, 5)
 #define EGRESS_IP3		IPV4(3, 3, 4, 5)
 
+#include "egressgw_policy.h"
+
 static volatile const __u8 *client_mac  = mac_one;
 static volatile const __u8 *gateway_mac = mac_two;
 static volatile const __u8 *ext_svc_mac = mac_three;
@@ -44,36 +46,6 @@ static __always_inline __be16 client_port(__u16 t)
 {
 	return CLIENT_PORT + bpf_htons(t);
 }
-
-#ifdef ENABLE_EGRESS_GATEWAY
-static __always_inline void add_egressgw_policy_entry(__be32 saddr, __be32 daddr, __u8 cidr,
-						      __be32 gateway_ip, __be32 egress_ip)
-{
-	struct egress_gw_policy_key in_key = {
-		.lpm_key = { EGRESS_PREFIX_LEN(cidr), {} },
-		.saddr   = saddr,
-		.daddr   = daddr,
-	};
-
-	struct egress_gw_policy_entry in_val = {
-		.egress_ip  = egress_ip,
-		.gateway_ip = gateway_ip,
-	};
-
-	map_update_elem(&EGRESS_POLICY_MAP, &in_key, &in_val, 0);
-}
-
-static __always_inline void del_egressgw_policy_entry(__be32 saddr, __be32 daddr, __u8 cidr)
-{
-	struct egress_gw_policy_key in_key = {
-		.lpm_key = { EGRESS_PREFIX_LEN(cidr), {} },
-		.saddr   = saddr,
-		.daddr   = daddr,
-	};
-
-	map_delete_elem(&EGRESS_POLICY_MAP, &in_key);
-}
-#endif /* ENABLE_EGRESS_GATEWAY */
 
 static __always_inline int egressgw_pktgen(struct __ctx_buff *ctx,
 					   struct egressgw_test_ctx test_ctx)
